+++ pdf-chatbot/requirements.txt
@@ -1,9 +1,8 @@
 faiss-cpu==1.7.4
+InstructorEmbedding==1.0.1
 gradio==4.15.0
-llama-index==0.10.19
-llama-index-embeddings-huggingface==0.1.4
-llama-index-llms-huggingface==0.1.4
-llama-index-llms-vllm==0.1.6
-llama-index-vector-stores-faiss==0.1.2
-vllm==0.3.3
+huggingface-hub==0.20.3
+langchain==0.1.3
+PyPDF2==3.0.1
+sentence-transformers==2.2.2
 transformers==4.37.0
